requirements.txt
python-dotenv
sentence_transformers
langchain_community
langchain
pypdf
langchain-text-splitters
transformers
groq 